+++ lib/forest_liana/bootstraper.rb
@@ -150,16 +150,20 @@ module ForestLiana
             request['forest-secret-key'] = ForestLiana.env_secret
             response = client.request(request)
 
-            # NOTICE: HTTP 404 Error
-            if response.is_a?(Net::HTTPNotFound)
+            unless response.body.blank?
+              warning = JSON.parse(response.body)['warning']
+            end
+
+            if response.is_a?(Net::HTTPNotFound) # NOTICE: HTTP 404 Error
               FOREST_LOGGER.error "Cannot find the project related to the " \
                 "env_secret you configured. Can you check on Forest that you " \
                 "copied it properly in the forest_liana initializer?"
-            # NOTICE: HTTP 400 Error
-            elsif response.is_a?(Net::HTTPBadRequest)
+            elsif response.is_a?(Net::HTTPBadRequest) # NOTICE: HTTP 400 Error
               FOREST_LOGGER.error "An error occured with the apimap sent to " \
                 "Forest. Please contact support@forestadmin.com for further " \
                 "investigations."
+            elsif warning
+              FOREST_LOGGER.warn warning
             end
           end
         rescue Errno::ECONNREFUSED